+++ HISTORY.md
@@ -51,6 +51,7 @@
 
 ### 2022-05-25
 * threads: update programmatic custom block-type changes in data references
+* byob: update manual custom block-type changes in data references
 
 ### 2022-05-23
 * blocks: fixed block label color when expanding or inserting variadic infix slots